refactor: split project health JS into a separate actions module
- extract generate/export click handlers from ai_helper_project_health.js
  into a new ai_helper_project_health_actions.js to stay under the
  lowered max-lines-per-function ESLint limit
- register the new script in all views that load the project health JS
- extract stuffTodo modal helpers and typo checker tooltip/button
  builders into standalone functions for the same reason
- lower max-lines-per-function from 130 to 90 and add ESLint globals
  for the new cross-file function references
- move the corresponding characterization tests to a new test file
