+++ app/views/ai_helper/shared/_html_header.html.erb
@@ -23,6 +23,7 @@
     <%= javascript_include_tag('typo_checker/ai_helper_typo_checker.js', plugin: :redmine_ai_helper) %>
     <%= javascript_include_tag('typo_checker/ai_helper_typo_suggestion_overlay.js', plugin: :redmine_ai_helper) %>
     <%= javascript_include_tag('project_health/ai_helper_project_health.js', plugin: :redmine_ai_helper) %>
+    <%= javascript_include_tag('project_health/ai_helper_project_health_actions.js', plugin: :redmine_ai_helper) %>
     <%= javascript_include_tag('autocompletion/ai_helper_assignment_suggestion.js', plugin: :redmine_ai_helper) %>
     <%= javascript_include_tag('project_health/ai_helper_master_detail.js', plugin: :redmine_ai_helper) %>
     <%= javascript_include_tag('sub_issues/ai_helper_sub_issues.js', plugin: :redmine_ai_helper) %>
