+++ app/views/ai_helper_dashboard/index.html.erb
@@ -5,6 +5,7 @@
 <% content_for :header_tags do %>
   <%= javascript_include_tag 'shared/ai_helper_markdown_parser', plugin: 'redmine_ai_helper' %>
   <%= javascript_include_tag 'project_health/ai_helper_project_health', plugin: 'redmine_ai_helper' %>
+  <%= javascript_include_tag 'project_health/ai_helper_project_health_actions', plugin: 'redmine_ai_helper' %>
   <%= javascript_include_tag 'project_health/ai_helper_master_detail', plugin: 'redmine_ai_helper' %>
   <%= stylesheet_link_tag 'ai_helper_rtl', plugin: 'redmine_ai_helper' if l(:direction) == 'rtl' %>
 <% end %>
