+++ assets/javascripts/stuff_todo/ai_helper_stuff_todo.js
@@ -2,80 +2,18 @@
 if (!window.aiHelperStuffTodoInitialized) {
   window.aiHelperStuffTodoInitialized = true;
 
-document.addEventListener('DOMContentLoaded', function() {
-
-  // Retrieve configuration from meta tags
-  const urlMeta = document.querySelector('meta[name="ai-helper-stuff-todo-url"]');
-  const errorMeta = document.querySelector('meta[name="ai-helper-stuff-todo-error"]');
-
-  // Hide the menu link on non-project pages where meta tag is not present
-  const menuLink = document.getElementById('ai-helper-stuff-todo-link');
-  if (!urlMeta) {
-    if (menuLink) {
-      menuLink.closest('li').style.display = 'none';
-    }
-    return;
-  }
-
-  const stuffTodoUrl = urlMeta.getAttribute('content');
-  // Show only the link element itself (minimal change).
-  // Avoid touching parent <li> or extra logic — just ensure the anchor is visible.
-  if (menuLink) {
-    menuLink.style.display = 'inline-block';
-  }
-
-  // Initialize markdown parser
-  let parser;
-  try {
-    if (typeof AiHelperMarkdownParser !== 'undefined') {
-      parser = new AiHelperMarkdownParser();
-    } else {
-      return;
-    }
-  } catch {
-    return;
-  }
-
-  // Get modal elements from server-rendered HTML (ERB template)
-  const overlay = document.getElementById('ai-helper-stuff-todo-overlay');
-  const modal = document.getElementById('ai-helper-stuff-todo-modal');
-  const closeBtn = document.getElementById('ai-helper-stuff-todo-close');
-  const body = document.getElementById('ai-helper-stuff-todo-body');
-
-  // Exit if modal elements are not found (should not happen if template is rendered correctly)
-  if (!overlay || !modal || !closeBtn || !body) {
-    return;
-  }
-
   let currentEventSource = null;
 
   /**
-   * Open the stuff-todo modal and start streaming suggestions into it.
-   */
-  function openModal() {
-    overlay.style.display = 'block';
-    modal.style.display = 'block';
-    body.innerHTML = '<div class="ai-helper-loader"></div>';
-    streamStuffTodo();
-  }
-
-  /**
-   * Close the stuff-todo modal and abort any in-flight streaming.
-   */
-  function closeModal() {
-    overlay.style.display = 'none';
-    modal.style.display = 'none';
-    if (currentEventSource) {
-      currentEventSource.close();
-      currentEventSource = null;
-    }
-  }
-
-  /**
    * Open an SSE connection to `stuffTodoUrl` and render the streamed
-   * markdown content into the modal body as it arrives.
+   * markdown content into `body` as it arrives.
+   * @param {string} stuffTodoUrl - SSE endpoint to stream suggestions from.
+   * @param {Element|null} errorMeta - Meta tag holding the localized error message.
+   * @param {object} parser - Markdown parser used to render streamed content.
+   * @param {HTMLElement} body - Modal body element to render content into.
+   * @returns {void}
    */
-  function streamStuffTodo() {
+  function streamStuffTodo(stuffTodoUrl, errorMeta, parser, body) {
     if (currentEventSource) {
       currentEventSource.close();
       currentEventSource = null;
@@ -127,28 +65,105 @@ document.addEventListener('DOMContentLoaded', function() {
     };
   }
 
-  // Event handler for the menu link added by Redmine's MenuManager
-  if (menuLink) {
-    menuLink.addEventListener('click', function(e) {
-      e.preventDefault();
-      openModal();
-    });
+  /**
+   * Open the stuff-todo modal and start streaming suggestions into it.
+   * @param {HTMLElement} overlay - Modal overlay element to show.
+   * @param {HTMLElement} modal - Modal element to show.
+   * @param {HTMLElement} body - Modal body element to reset and stream content into.
+   * @param {string} stuffTodoUrl - SSE endpoint to stream suggestions from.
+   * @param {Element|null} errorMeta - Meta tag holding the localized error message.
+   * @param {object} parser - Markdown parser used to render streamed content.
+   * @returns {void}
+   */
+  function openStuffTodoModal(overlay, modal, body, stuffTodoUrl, errorMeta, parser) {
+    overlay.style.display = 'block';
+    modal.style.display = 'block';
+    body.innerHTML = '<div class="ai-helper-loader"></div>';
+    streamStuffTodo(stuffTodoUrl, errorMeta, parser, body);
   }
 
-  closeBtn.addEventListener('click', function() {
-    closeModal();
-  });
-
-  overlay.addEventListener('click', function() {
-    closeModal();
-  });
-
-  // Close on Escape key
-  document.addEventListener('keydown', function(e) {
-    if (e.key === 'Escape' && modal.style.display === 'block') {
-      closeModal();
+  /**
+   * Close the stuff-todo modal and abort any in-flight streaming.
+   * @param {HTMLElement} overlay - Modal overlay element to hide.
+   * @param {HTMLElement} modal - Modal element to hide.
+   * @returns {void}
+   */
+  function closeStuffTodoModal(overlay, modal) {
+    overlay.style.display = 'none';
+    modal.style.display = 'none';
+    if (currentEventSource) {
+      currentEventSource.close();
+      currentEventSource = null;
     }
+  }
+
+  document.addEventListener('DOMContentLoaded', function() {
+
+    // Retrieve configuration from meta tags
+    const urlMeta = document.querySelector('meta[name="ai-helper-stuff-todo-url"]');
+    const errorMeta = document.querySelector('meta[name="ai-helper-stuff-todo-error"]');
+
+    // Hide the menu link on non-project pages where meta tag is not present
+    const menuLink = document.getElementById('ai-helper-stuff-todo-link');
+    if (!urlMeta) {
+      if (menuLink) {
+        menuLink.closest('li').style.display = 'none';
+      }
+      return;
+    }
+
+    const stuffTodoUrl = urlMeta.getAttribute('content');
+    // Show only the link element itself (minimal change).
+    // Avoid touching parent <li> or extra logic — just ensure the anchor is visible.
+    if (menuLink) {
+      menuLink.style.display = 'inline-block';
+    }
+
+    // Initialize markdown parser
+    let parser;
+    try {
+      if (typeof AiHelperMarkdownParser !== 'undefined') {
+        parser = new AiHelperMarkdownParser();
+      } else {
+        return;
+      }
+    } catch {
+      return;
+    }
+
+    // Get modal elements from server-rendered HTML (ERB template)
+    const overlay = document.getElementById('ai-helper-stuff-todo-overlay');
+    const modal = document.getElementById('ai-helper-stuff-todo-modal');
+    const closeBtn = document.getElementById('ai-helper-stuff-todo-close');
+    const body = document.getElementById('ai-helper-stuff-todo-body');
+
+    // Exit if modal elements are not found (should not happen if template is rendered correctly)
+    if (!overlay || !modal || !closeBtn || !body) {
+      return;
+    }
+
+    // Event handler for the menu link added by Redmine's MenuManager
+    if (menuLink) {
+      menuLink.addEventListener('click', function(e) {
+        e.preventDefault();
+        openStuffTodoModal(overlay, modal, body, stuffTodoUrl, errorMeta, parser);
+      });
+    }
+
+    closeBtn.addEventListener('click', function() {
+      closeStuffTodoModal(overlay, modal);
+    });
+
+    overlay.addEventListener('click', function() {
+      closeStuffTodoModal(overlay, modal);
+    });
+
+    // Close on Escape key
+    document.addEventListener('keydown', function(e) {
+      if (e.key === 'Escape' && modal.style.display === 'block') {
+        closeStuffTodoModal(overlay, modal);
+      }
+    });
   });
-});
 
 } // End guard against multiple script loading
